% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kcop-package.R
\docType{package}
\name{Kcop-package}
\alias{Kcop}
\alias{Kcop-package}
\title{Kcop: Smooth Test for Equality of Copulas and Clustering Multivariate}
\description{
Implements approaches of non-parametric smooth test to 
            compare simultaneously K(K>1) copulas and non-parametric clustering 
            of multivariate populations with arbitrary sizes. 
            See Yves I. Ngounou Bakam and Denys Pommeret (2022) <arXiv:2112.05623> and
            Yves I. Ngounou Bakam and Denys Pommeret (2022) <arXiv:2211.06338>.
}
\author{
\strong{Maintainer}: Yves Ismael Ngounou Bakam \email{yvesngounou20@gmail.com}

Authors:
\itemize{
  \item Denys Pommeret
}

}
\keyword{internal}
